Trabalhos/Jorge_analise_combinatoria/requirements2.txt
et_xmlfile==2.0.0
numpy==2.1.3
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
setuptools==75.5.0
six==1.16.0
style==1.1.0
tzdata==2024.2
update==0.0.1
